Moved lots of stuff, broke everything - Using yaml to configure music box
